:recycle: refactor: default theme
